requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.11
djangorestframework==3.15.1
django-filter==24.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Pillow==10.0.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
